refactor(foundation-demo): move sidebar to its own component
move sidebar to its own component
