+++ apps/foundation-demo/src/app/misc/sidebar/components/sidebar/sidebar.component.scss
