+++ apps/foundation-demo/src/app/misc/sidebar/reducers/sidebar.reducer.ts
